refactor: fixed card list test to work with new props
