+++ src/_test_/card-list.test.tsx
@@ -1,18 +1,32 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 import { CardList } from '../components/card-list/card-list';
+import type { CardListProps } from '../types/types';
+
+const defaultCardListProps: CardListProps = {
+  animalsList: [],
+  loading: false,
+  error: null,
+  onNextPage: () => {},
+  onPrevPage: () => {},
+  currentPage: 0,
+};
 
 describe('CardList component', () => {
   // Shows spinner when loading is true
   it('Shows loading state while fetching data', () => {
-    render(<CardList animalsList={[]} loading={true} error={null} />);
+    const testProps = {
+      ...defaultCardListProps,
+      loading: true,
+    };
+    render(<CardList {...testProps} />);
     const spinner = screen.getByTestId('spinner');
     expect(spinner).toBeInTheDocument();
   });
 
   // Shows message when no data is available
   it('Displays "no results" message when data array is empty', () => {
-    render(<CardList animalsList={[]} loading={false} error={null} />);
+    render(<CardList {...defaultCardListProps} />);
 
     const message = screen.getByText(/no animals found/i);
     expect(message).toBeInTheDocument();
@@ -41,8 +55,11 @@ describe('CardList component', () => {
       },
     ];
 
-    render(<CardList animalsList={mockAnimals} loading={false} error={null} />);
-
+    const testProps = {
+      ...defaultCardListProps,
+      animals: mockAnimals,
+    };
+    render(<CardList {...testProps} />);
     const items = screen.getAllByRole('heading', { level: 3 });
     expect(items.length).toBe(2);
   });
